requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
matplotlib>=3.8.0
moviepy>=1.0.3
Pillow>=10.0.0
numpy>=1.26.0
